When running some tests on unsupported OS (Ubuntu), the confusing failure messages arise caused by the fact that some variables are not initialized in setUp but used in the tearDown.
In memory/pmem* and ndctl, let's check if we have corresponding attributes set
before trying to use them.

In selinux-tests, we can initiate the problematic variable to None in the beginning
and then use it only if it is not empty.

ras/lparstat is just missing explicit test cancelation for the OS other than
rhel/suse.

+++ memory/ndctl.py
@@ -1016,15 +1016,16 @@ class NdctlTest(Test):
 
     @avocado.fail_on(pmem.PMemException)
     def tearDown(self):
-        if self.part:
+        if hasattr(self, 'part') and self.part:
             self.part.unmount()
-        if self.disk:
+        if hasattr(self, 'disk') and self.disk:
             self.log.info("Removing the FS meta created on %s", self.disk)
             delete_fs = "dd if=/dev/zero bs=1M count=1024 of=%s" % self.disk
             if process.system(delete_fs, shell=True, ignore_status=True):
                 self.fail("Failed to delete filesystem on %s" % self.disk)
 
-        if not self.preserve_setup:
-            if self.plib.run_ndctl_list('-N'):
-                self.plib.destroy_namespace(force=True)
-            self.plib.disable_region()
+        if hasattr(self, 'preserve_setup') and not self.preserve_setup:
+            if hasattr(self, 'plib'):
+                if self.plib.run_ndctl_list('-N'):
+                    self.plib.destroy_namespace(force=True)
+                self.plib.disable_region()
